--- inprovement.md
@@ -1,8 +1,0 @@
-// get ll the quizzes just hit the get API
-// http://localhost:8000/api/v1/user/quizzes
-
-// get quetsion with their option
-// http://localhost:8000/api/v1/quizze/questions/Quiz_id
-
-// example
-//http://localhost:8000/api/v1/quizze/questions/a1b2c3d4-e5f6-7a89-b0c1-d2e3f4a5b6c7
