+++ src/app.js
@@ -32,7 +32,7 @@ import examAndQuestRouter from "#routes/exam.route.js";
 import healthcheckRouter from "#routes/healthcheck.route.js";
 
 // routes declaration
-app.use("/v1", getAllCollege);
+app.use("/v1/colleges", getAllCollege);
 app.use("/v1/users", userRouter);
 app.use("/v1/exams", examAndQuestRouter);
 app.use("/v1/checks", healthcheckRouter);
