+++ src/routes/exam.route.js
@@ -1,11 +1,11 @@
 import { Router } from "express";
 import { verifyJWT } from "#middleware/auth.middleware.js";
 
-import { inviteStudent } from "#controllers/enrolled.controller.js";
 import { createExam } from "#controllers/exams/createExam.controller.js";
 import { getExams } from "#controllers/exams/getExams.controller.js";
 import { createQuestion } from "#controllers/exams/createQuestion.controller.js";
 import { getQuestions } from "#controllers/exams/getQuestions.controller.js";
+import { inviteStudent } from "#controllers/exams/inviteStudents.controller.js";
 
 const router = Router();
